requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
openai==2.16.0
packaging==26.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pytz==2025.2
PyYAML==6.0.3
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
